move localexec to common/shell-local dir since it is related to shelling out locally
