+++ common/step_create_cdrom.go
@@ -11,7 +11,7 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/hashicorp/packer/helper/builder/localexec"
+	"github.com/hashicorp/packer/common/shell-local/localexec"
 	"github.com/hashicorp/packer/helper/multistep"
 	"github.com/hashicorp/packer/packer"
 	"github.com/hashicorp/packer/packer/tmp"
